+++ CHANGELOG.md
@@ -5,7 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]## [0.5.4.3b1]
+### Changed
 - Fixed indentation bug in domodeldeployer
+### Added
+- DOModelDeployer support for gz/zip package files
 
 ## [Unreleased]## [0.5.4.3b0]
 ### Changed